convert send to deprecated-signatures
