+++ transforms/__test__/deprecated-signatures.spec.ts
@@ -1,0 +1,3 @@
+import { testSpecBuilder } from './util'
+
+testSpecBuilder('deprecated-signatures')
